Performs the logic for the card handler and changing cards in the different decks.
Unfisnished/untested, was not able to fully implement the Ficher Yates shuffle due to time constrains although other mode should be working.
Depending n the mode the game is set in, moves cards around in the 3 different decks, or adds new cards to the discard durring buy phase/game start.